Update import statements to include file extensions
